Divide up interop rpc module into submodules
